Use tpl function for defaultTags in Deployment (#97)
This PR modifies the `defaultTags` argument in the Deployment
configuration. Previously, the `defaultTags` value was directly
converted to JSON. This PR changes this behavior by first passing the
`defaultTags` through the Helm `tpl` function before converting it to
JSON.

The `tpl` function allows us to use templating in the
`defaultTags`values. This means we can now use values from other Helm
variables within `defaultTags`. This provides more flexibility and
dynamic configuration options for users.

This change is backward compatible. If no templating is used in
defaultTags, the behavior remains the same as before.

Co-authored-by: jallphin <jallphin@adobe.com>